+++ container_writing.py
